Board tracks ValidOrders, cyclic dep workaround
Previously we had

  Board.hs
  ResolvedOrder.hs
  OrderInvalid.hs

with cyclic dependencies: Board and OrderInvalid are mutually dependent;
validity depends upon a Board, so we need its type, and a Board uses
order validity, so we need types and functions; we can only resolve
valid orders, so ResolvedOrder needs OrderInvalid, and of course Board
is interested in the ResolvedOrder type, so we have something like

  Board  <-------  OrderInvalid
    |    ------->      ^
    |                  |
    v                  |
  ResolvedOrder -------+

where a --> b means a depends upon b.

I tried to resolve this using hs-boot files, but I found that the
Board.hs-boot required ResolvedOrder.hs, which in turn requires
OrderInvalid, and therefore requires Board.hs-boot again! A cycle even
in the hs-boot files.

Discourages, I decided on the solution implemented here: merge
OrderInvalid and ResolvedOrder.

This commit also includes big changes to Board.hs: every order in the
Board is guaranteed to be valid.